The Gaya browser does not allow string manipulation on strings that contain square brackets, which was messing up the advanced JS version of the skins when viewing data from a NAS.
I have changed this to disable the JS editing functions for most NAS data and optimised the user experience as much as I can until such a time as the show() method provides the episode title as just a title and the episode date as a separate parameter. At this point the advanced / basic split will also no longer be required.